# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dataplat API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sources/collections": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Collections",
        "description": "List source collections owned by the authenticated user.\n\nReturns a paginated list scoped to the caller's owner_id.\n`total` is the count of ALL the caller's collections (not just the current\npage), enabling the client to compute page counts without a second request.\n\nPagination:\n  limit  \u2014 max items per page (1\u2013200, default 20)\n  offset \u2014 number of rows to skip (default 0)\n\nOrdering: id ASC (oldest first, stable).\nAuth required (F-008).",
        "operationId": "list_collections_api_sources_collections_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Create Source Collection",
        "description": "Create a new source collection.\n\nCreates a source_collection row in Postgres and returns the new record.\nDuplicate names return 409 (detected via the Postgres UNIQUE constraint\n`source_collection_name_key`). Auth required (F-008).",
        "operationId": "create_collection_api_sources_collections_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCollectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceCollectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sources/collections/{id}/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Sources in Collection",
        "description": "Return a paginated list of sources belonging to the specified collection.\n\nOwner-scoping (F-014 agreed.md \u00a73.5):\n  Step 1 \u2014 Verify the collection exists AND is owned by the caller.\n           If it does not exist or belongs to another user \u2192 404.\n           Returning 404 (not 403) for both cases prevents enumeration leaks.\n  Step 2 \u2014 List sources in that collection (paginated page + total count).\n\nThe ownership check is a dedicated query that short-circuits to 404 before\nthe paginated source queries are issued.  Embedding ownership as a JOIN on\nthe source queries would silently return {\"items\":[], \"total\":0} (HTTP 200)\nfor an unowned collection \u2014 violating the no-enumeration-leak invariant.\n\nPagination:\n  limit  \u2014 max items per page (1\u2013200, default 20)\n  offset \u2014 number of rows to skip (default 0)\ntotal    \u2014 count of ALL sources in the collection (not just the current page)\nOrdering: Source.id ASC (oldest first, stable).\n\nAuth required (F-008).",
        "operationId": "list_sources_by_collection_api_sources_collections__id__sources_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sources/upload": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Upload PDF Source",
        "description": "Upload a PDF file as a new source.\n\nStores the file in MinIO at s3://sources/{source_id}/original.pdf and\nwrites a source row to Postgres with sha256, storage_uri, kind='file',\nmime_type='application/pdf'.  Returns the new source id and storage_uri.\n\nAfter commit, notifies Dagster on a best-effort basis (F-012):\n  1. addDynamicPartition: registers src_{id} in the \"sources\" partition def.\n  2. reportRunlessAssetEvents: records a materialization event for 'source'.\nIf either Dagster call fails, the upload still returns 201 \u2014 the source is\ndurable in DB+MinIO regardless. A Dagster outage does not roll back the upload.\n\nAtomicity (agreed.md \u00a73-D3):\n  flush (get id) \u2192 set uri/partition_key \u2192 S3 upload \u2192 commit.\n  If S3 upload fails, the exception propagates; the open DB transaction is\n  implicitly rolled back when the connection returns to the pool (the\n  get_session() context manager calls session.close(), NOT rollback()).\n  Do NOT add explicit session.rollback() here \u2014 let exceptions propagate.\n  Known acceptable leak: if commit() fails after a successful S3 upload,\n  the MinIO object persists with no corresponding DB row (agreed.md \u00a73-D6).\n\nAuth required (F-008).",
        "operationId": "upload_source_api_sources_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_source_api_sources_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/sources/{source_id}/documents": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Document Variants",
        "description": "Return a flat list of all document_variant rows for the given source.\n\nOwner-scoping (F-020 agreed.md \u00a74.2):\n  Step 1 \u2014 Verify the source exists AND is accessible to the caller.\n           A source is accessible if it has no collection (collection_id IS NULL)\n           or its collection is owned by the caller.\n           If the source does not exist or belongs to another user's collection\n           \u2192 404 (prevents enumeration leaks, same as GET /{id}).\n  Step 2 \u2014 Fetch all document_variant rows for the source, ordered by id ASC.\n           Returns an empty list ([]) when the source exists but has no\n           variants yet (i.e. extraction has not run).\n\nReturns a plain JSON array (not paginated) \u2014 variants per source are small\n(typically 1\u20133 in practice).\n\nAuth required (F-008).",
        "operationId": "list_document_variants_api_sources__source_id__documents_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentVariantRead"
                  },
                  "title": "Response List Document Variants Api Sources  Source Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_source_api_sources_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "collection_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_source_api_sources_upload_post"
      },
...
      "CollectionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceCollectionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "CollectionListResponse",
        "description": "Response for GET /api/sources/collections (F-010)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceCollectionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "dataset_card_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Card Md"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SourceCollectionCreate",
        "description": "Request body for POST /api/sources/collections."
      },
...
      "SourceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SourceListResponse",
        "description": "Response for GET /api/sources/collections/{id}/sources (F-014).\n\nPaginated list of source records belonging to a single collection.\nMirrors CollectionListResponse shape from F-010:\n  items \u2014 the current page of SourceRead records (ordered by id ASC)\n  total \u2014 count of ALL sources in the collection (not just the page)"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}